layers/utils_layer/requirements.txt
Pillow == 9.2.0  # for Python 3.9
